Nearly fully working cart

+++ Off-Black.Services/DTO/OrderDTO.cs
@@ -6,7 +6,7 @@ namespace Off_Black.Services.DTO
     public class OrderDTO
     {
         public int OrderID { get; set; }
-        public int TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public DateTime OrderDate { get; set; }
 
         public ICollection<OrderItemDTO> OrderItems { get; set; }
